add NoteViewModel
