+++ app/src/main/java/com/example/compnote/presentation/MainViewModel.kt
@@ -17,13 +17,9 @@ import javax.inject.Inject
 @HiltViewModel
 class MainViewModel @Inject constructor(
     private val noteReadAllUseCase: NoteReadAllUseCase,
-    private val noteDeleteByIdUseCase: NoteDeleteByIdUseCase,
-    private val noteUpdateUseCase: NoteUpdateUseCase,
-    private val noteGetByIdUseCase: NoteGetByIdUseCase
 ) : ViewModel() {
 
     val allListNote = MutableLiveData<List<Note>>()
-    val note = MutableLiveData<Note>()
 
     fun getAllNotes() {
         viewModelScope.launch(Dispatchers.IO) {
@@ -36,36 +32,4 @@ class MainViewModel @Inject constructor(
             }
         }
     }
-
-    fun getNoteById(id: Int) {
-        viewModelScope.launch(Dispatchers.IO) {
-            noteGetByIdUseCase.execute(id = id).collect { response ->
-                when (response) {
-                    is Response.Loading -> {}
-                    is Response.Fail -> {}
-                    is Response.Success -> this@MainViewModel.note.postValue(response.data)
-                }
-            }
-        }
-    }
-
-    fun updateNote(note: Note): Flow<Boolean> = flow {
-        noteUpdateUseCase.execute(note = note).collect { response ->
-            when (response) {
-                is Response.Loading -> {}
-                is Response.Fail -> emit(false)
-                is Response.Success -> emit(response.data)
-            }
-        }
-    }.flowOn(Dispatchers.IO)
-
-    fun deleteNoteById(id: Int): Flow<Boolean> = flow<Boolean> {
-        noteDeleteByIdUseCase.execute(id = id).collect { response ->
-            when (response) {
-                is Response.Loading -> {}
-                is Response.Fail -> emit(false)
-                is Response.Success -> emit(response.data)
-            }
-        }
-    }.flowOn(Dispatchers.IO)
 }
