+++ app/src/main/java/com/example/compnote/presentation/screens/add/AddScreen.kt
@@ -1,4 +1,4 @@
-package com.example.compnote.presentation.screens
+package com.example.compnote.presentation.screens.add
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -16,10 +16,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import com.example.compnote.presentation.MainViewModel
 import com.example.compnote.domain.models.Note
 import com.example.compnote.presentation.navigation.NavRoute
-import com.example.compnote.presentation.screens.add.AddViewModel
 import com.example.compnote.presentation.util.Constants
 import com.example.compnote.presentation.util.Constants.Keys.ADD_NOTE
 import kotlinx.coroutines.launch
